Revert "src: move *Exceptions out to separate cc/h"
This reverts commit 9349e15daa803fc626581802fa827afe8d00340e.
