+++ node.gyp
@@ -314,7 +314,6 @@
         'src/connection_wrap.cc',
         'src/connect_wrap.cc',
         'src/env.cc',
-        'src/exceptions.cc',
         'src/fs_event_wrap.cc',
         'src/handle_wrap.cc',
         'src/js_stream.cc',
@@ -377,7 +376,6 @@
         'src/connect_wrap.h',
         'src/env.h',
         'src/env-inl.h',
-        'src/exceptions.h',
         'src/handle_wrap.h',
         'src/js_stream.h',
         'src/module_wrap.h',
